+++ CHANGELOG.md
@@ -12,6 +12,10 @@ listed in the changelog.
 
 ## [0.6.0] - 2022-07-20
 
+### Changed
+
+- Always print Aqua log output ([#593](https://github.com/opendevstack/ods-pipeline/issues/593))
+
 ### Added
 
 - Enable build caching for gradle builds according to `docs/adr/20220314-caching-build-tasks.md`.